requirements.txt
torch
torchvision
numpy
pillow
fastapi
uvicorn
pydantic
python-multipart
tqdm
scikit-learn
pandas
